✅🚚♻️ Update and restructure testing for crossover functionality
The testing code for the crossover functionality has been substantially expanded, improving the coverage. Moved the testing functions `crossover should return a list of offspring` and `crossover should throw an exception on incorrect configuration` from `CrossoverTest.kt` to `Operators.kt`. These changes aim to improve code reusability and test coverage.
