+++ keen-core/src/test/kotlin/cl/ravenhill/keen/assertions/Operators.kt
@@ -1,19 +1,37 @@
 package cl.ravenhill.keen.assertions
 
 import cl.ravenhill.jakt.exceptions.CompositeException
+import cl.ravenhill.keen.Domain
 import cl.ravenhill.keen.arb.datatypes.arbProbability
+import cl.ravenhill.keen.arb.genetic.arbGenotype
+import cl.ravenhill.keen.arb.genetic.chromosomes.arbIntChromosome
+import cl.ravenhill.keen.arb.operators.arbBaseCrossover
 import cl.ravenhill.keen.assertions.should.shouldHaveInfringement
+import cl.ravenhill.keen.exceptions.CrossoverException
 import cl.ravenhill.keen.exceptions.MutatorConfigException
+import cl.ravenhill.keen.genetic.Genotype
+import cl.ravenhill.keen.genetic.chromosomes.Chromosome
+import cl.ravenhill.keen.genetic.chromosomes.numeric.IntChromosome
 import cl.ravenhill.keen.genetic.genes.Gene
+import cl.ravenhill.keen.genetic.genes.numeric.IntGene
+import cl.ravenhill.keen.operators.alteration.crossover.Crossover
 import cl.ravenhill.keen.operators.alteration.mutation.GeneMutator
 import cl.ravenhill.keen.operators.alteration.mutation.Mutator
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.freeSpec
+import io.kotest.core.spec.style.scopes.FreeSpecContainerScope
+import io.kotest.datatest.withData
 import io.kotest.matchers.shouldBe
 import io.kotest.property.Arb
 import io.kotest.property.arbitrary.double
+import io.kotest.property.arbitrary.filter
 import io.kotest.property.arbitrary.filterNot
+import io.kotest.property.arbitrary.flatMap
+import io.kotest.property.arbitrary.int
+import io.kotest.property.arbitrary.list
+import io.kotest.property.arbitrary.map
 import io.kotest.property.checkAll
+import kotlin.random.Random
 
 /**
  * Tests the individual rate property of a Mutator in a property-based testing style.
@@ -247,4 +265,124 @@ fun <T, G> `test Gene Mutator gene rate`(
             }
         }
     }
+}
+
+suspend fun FreeSpecContainerScope.`crossover should return a list of offspring`() {
+    "should return a list of offspring genotypes" - {
+        withData(
+            listOf(
+                Triple(
+                    genericCrossover(1),
+                    listOf(Genotype(IntChromosome(*arrayOf()))),
+                    listOf(Genotype(IntChromosome(*arrayOf())))
+                ),
+                Triple(
+                    genericCrossover(2),
+                    listOf(
+                        Genotype(IntChromosome(1, 2)),
+                        Genotype(IntChromosome(3, 4))
+                    ),
+                    listOf(
+                        Genotype(IntChromosome(3, 4)),
+                        Genotype(IntChromosome(1, 2))
+                    )
+                ),
+            )
+        ) { (crossover, parents, expected) ->
+            Domain.random = Random(420)
+            val offspring = crossover.crossover(parents)
+            offspring shouldBe expected
+        }
+    }
+}
+
+suspend fun FreeSpecContainerScope.`crossover should throw an exception on incorrect configuration`() {
+    "should throw an exception if" - {
+        "the number of parents is different from the inputs" {
+            checkAll(mismatchedCrossoverPairs()) { (crossover, parents) ->
+                shouldThrow<CompositeException> {
+                    crossover.crossover(parents)
+                }.shouldHaveInfringement<CrossoverException>(
+                    "The number of inputs (${parents.size}) must be equal to the number of parents " +
+                            "(${crossover.numParents})"
+                )
+            }
+        }
+
+        "the genotypes have different number of chromosomes" - {
+            withData(
+                genericCrossover() to listOf(
+                    Genotype(IntChromosome(IntGene(1))),
+                    Genotype(IntChromosome(IntGene(1)), IntChromosome(IntGene(1)))
+                ),
+                genericCrossover() to listOf(
+                    Genotype(IntChromosome(IntGene(1)), IntChromosome(IntGene(1))),
+                    Genotype(IntChromosome(IntGene(1)))
+                ),
+                genericCrossover() to listOf(
+                    Genotype(IntChromosome(IntGene(1)), IntChromosome(IntGene(1))),
+                    Genotype(IntChromosome(IntGene(1)), IntChromosome(IntGene(1)), IntChromosome(IntGene(1)))
+                )
+            ) { (crossover, parents) ->
+                shouldThrow<CompositeException> {
+                    crossover.crossover(parents)
+                }.shouldHaveInfringement<CrossoverException>(
+                    "Genotypes must have the same number of chromosomes"
+                )
+            }
+        }
+
+        "any parent genotype is empty" - {
+            withData(
+                Triple(
+                    genericCrossover(),
+                    listOf(
+                        Genotype(),
+                        Genotype(IntChromosome(IntGene(1)))
+                    ),
+                    listOf(0)
+                ),
+                Triple(
+                    genericCrossover(),
+                    listOf(
+                        Genotype(IntChromosome(IntGene(1))),
+                        Genotype()
+                    ),
+                    listOf(1)
+                ),
+                Triple(
+                    genericCrossover(),
+                    listOf(
+                        Genotype(),
+                        Genotype()
+                    ),
+                    listOf(0, 1)
+                )
+            ) { (crossover, parents, emptyIndex) ->
+                val ex = shouldThrow<CompositeException> {
+                    crossover.crossover(parents)
+                }
+                emptyIndex.forEach { index ->
+                    ex.shouldHaveInfringement<CrossoverException>(
+                        "The number of chromosomes in parent $index must be greater than 0"
+                    )
+                }
+            }
+        }
+    }
+}
+
+fun genericCrossover(numParents: Int = 2): Crossover<Int, IntGene> = object : Crossover<Int, IntGene> {
+    override val numOffspring: Int = 2
+    override val numParents: Int = numParents
+    override val chromosomeRate: Double = 0.5
+    override val exclusivity: Boolean = false
+
+    override fun crossoverChromosomes(chromosomes: List<Chromosome<Int, IntGene>>) = chromosomes.reversed()
+}
+
+private fun mismatchedCrossoverPairs() = arbBaseCrossover<Int, IntGene>().flatMap { crossover ->
+    Arb.list(arbGenotype(arbIntChromosome(Arb.int(5..5)))).filter {
+        it.size != crossover.numParents
+    }.map { parents -> crossover to parents }
 }
